Table VersemntCsg

+++ database/migrations/2024_11_29_135611_create_versement_csgs_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->unsignedBigInteger('code_vcsg')->autoIncrement();
             $table->string('libelle', 50);
             $table->date('date');
-            $table->smallInteger('somme');
+            $table->Integer('somme');
             $table->unsignedBigInteger('verseur');
             $table->unsignedBigInteger('encaisseur');
             $table->foreign('verseur')->references('id')->on('users')->onDelete('cascade');
